Remove some duplicated or outdated components
