+++ app/(noMap)/about/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from "next"
 
-import { WagtailBlock } from "Features/WagtailApi/WagtailBlock.next"
+import { WagtailBlock } from "Features/WagtailApi/WagtailBlock"
 
 export default function About() {
   return <WagtailBlock pageId="5" />
