+++ app/(platformMap)/@sidebar/page.tsx
@@ -1,4 +1,4 @@
-import { WagtailBlock } from "Features/WagtailApi/WagtailBlock.next"
+import { WagtailBlock } from "Features/WagtailApi/WagtailBlock"
 
 export default function HomeSidebar() {
   return <WagtailBlock pageId="4" />
