+++ src/Shared/google-analytics.tsx
@@ -1,3 +1,4 @@
+"use client"
 /**
  * Send Google analytics tracking info
  *
@@ -6,24 +7,22 @@
 
 import React, { useEffect } from "react"
 import ReactGA from "react-ga4"
-import { useLocation } from "react-router-dom"
+import { usePathname } from "next/navigation"
 
 interface Props {
   children: JSX.Element
   trackingId?: string
 }
 
-const GAListener: React.FC<Props> = ({ children, trackingId }: Props) => {
-  const location = useLocation()
+export function GAListener({ children, trackingId }: Props) {
+  const path = usePathname()
 
   useEffect(() => {
     if (trackingId) {
       ReactGA.initialize(trackingId)
-      ReactGA.send({ hitType: "pageview", page: `${location.pathname}${location.search}` })
+      ReactGA.send({ hitType: "pageview", page: `${path}` })
     }
-  }, [location, trackingId])
+  }, [path, trackingId])
 
   return children
 }
-
-export default GAListener
