--- src/queryClient.ts
@@ -1,10 +1,0 @@
-import { QueryClient } from "@tanstack/react-query"
-
-export const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      refetchOnWindowFocus: false,
-      // staleTime: 1000 * 60 * 5, // milliseconds
-    },
-  },
-})
